feat: Switch back to float from double for embedded target. Fix bug in polygon coloring when in color mode (#1)

+++ components/engine/include/camera.hpp
@@ -6,7 +6,7 @@
 class Camera {
 
 private:
-  double theta, phi;
+  float theta, phi;
   Vector3D forward;
   Vector3D up;
   Vector3D right;
@@ -22,25 +22,25 @@ public:
 
   void Translate(const Vector3D &v);
 
-  void Rotate(const double _t, const double _p);
+  void Rotate(const float _t, const float _p);
 
   Matrix GetWorldToCamera();
 
-  double GetPhi() const { return phi; }
-  double GetTheta() const { return theta; }
-  void SetAngles(const double _t, const double _p);
+  float GetPhi() const { return phi; }
+  float GetTheta() const { return theta; }
+  void SetAngles(const float _t, const float _p);
 
   Point3D GetPosition() const;
-  void SetPosition(const double x, const double y, const double z);
+  void SetPosition(const float x, const float y, const float z);
 
   Point3D GetForward() const;
-  void SetForward(const double x, const double y, const double z);
+  void SetForward(const float x, const float y, const float z);
 
   Point3D GetUp() const;
-  void SetUp(const double x, const double y, const double z);
+  void SetUp(const float x, const float y, const float z);
 
   Point3D GetRight() const;
-  void SetRight(const double x, const double y, const double z);
+  void SetRight(const float x, const float y, const float z);
 
   bool operator==(const Camera &c) {
     if (phi != c.GetPhi())
