+++ components/engine/include/vertex.hpp
@@ -14,14 +14,14 @@ class Vertex {
 public:
   union {
     struct {
-      double x, y, z, w; // 3D homogeneous coords ( get transformed into pixel space )
-      double ex, ey, ez; // 3D homogenous eye/camera coords ( w for this system is hw )
-      double u, v;       // 2D texture coords (range = [0,1]) ( w for this system is hw )
-      double r, g, b;    // RGB color (range = [0,1]), ( w for this system is hw )
-      double nx, ny, nz; // components of normal vector ( w for this system is hw )
-      double hw;         // homogeneous coordinate for interpolated values ( 1/w )
+      float x, y, z, w; // 3D homogeneous coords ( get transformed into pixel space )
+      float ex, ey, ez; // 3D homogenous eye/camera coords ( w for this system is hw )
+      float u, v;       // 2D texture coords (range = [0,1]) ( w for this system is hw )
+      float r, g, b;    // RGB color (range = [0,1]), ( w for this system is hw )
+      float nx, ny, nz; // components of normal vector ( w for this system is hw )
+      float hw;         // homogeneous coordinate for interpolated values ( 1/w )
     };
-    double data[NUM_VERTEX_DATA];
+    float data[NUM_VERTEX_DATA];
   };
 
   Vertex() {
@@ -29,8 +29,8 @@ public:
       data[i] = 0;
     w = hw = 1;
   }
-  Vertex(const double _x, const double _y, const double _z, const double _w = 1,
-         const double _u = 0, const double _v = 0) {
+  Vertex(const float _x, const float _y, const float _z, const float _w = 1, const float _u = 0,
+         const float _v = 0) {
     x = _x;
     y = _y;
     z = _z;
@@ -47,21 +47,21 @@ public:
   void Print() {}
 
   // Variable Setting Methods
-  void SetColor(const double _r, const double _g, const double _b) {
+  void SetColor(const float _r, const float _g, const float _b) {
     r = _r;
     g = _g;
     b = _b;
   }
-  void SetTextureCoords(const double _u, const double _v) {
+  void SetTextureCoords(const float _u, const float _v) {
     u = _u;
     v = _v;
   }
-  void SetEyeCoords(const double _x, const double _y, const double _z) {
+  void SetEyeCoords(const float _x, const float _y, const float _z) {
     ex = _x;
     ey = _y;
     ez = _z;
   }
-  void SetNormalCoords(const double _x, const double _y, const double _z) {
+  void SetNormalCoords(const float _x, const float _y, const float _z) {
     nx = _x;
     ny = _y;
     nz = _z;
@@ -70,7 +70,7 @@ public:
   // General Transformation Methods, only operate on x,y,z,w
   void Transform(const Matrix &_m);
   void Translate(const Vector3D &_v);
-  void Translate(const double _x, const double _y, const double _z);
+  void Translate(const float _x, const float _y, const float _z);
 
   // Pipeline Transformation Methods
   void TransformToCamera(const Matrix &_m);
@@ -81,10 +81,10 @@ public:
   Vertex &operator=(const Vertex &rhs);
   bool operator==(const Vertex &rhs) const;
   bool operator!=(const Vertex &rhs) const;
-  double &operator[](const int i) { return data[i]; }
-  double operator[](const int i) const { return data[i]; }
+  float &operator[](const int i) { return data[i]; }
+  float operator[](const int i) const { return data[i]; }
   Vertex operator-(const Vertex &rhs) const;
   Vertex operator+(const Vertex &rhs) const;
-  Vertex operator*(const double rhs) const;
-  Vertex operator/(const double rhs) const;
+  Vertex operator*(const float rhs) const;
+  Vertex operator/(const float rhs) const;
 };
